Split out the logic for the form into a separate component
